frontend/requirements.txt
requests==2.31.0
streamlit==1.32.2
pathlib[standard]
pandas[standard]
Pillow[standard]